#4 Create workaround for log4j2 config file
Rather than add a configuration file to the classpath, which has
proved to not be able to work, a ConfigurationFactory was created
instead. This allows the loggers to be configured programmatically.
The factory is set to initialize a configuration just like defined in
the log4j2.xml config. Along with this, InfrequentLogger was renamed to
DelayableLogger (this name is still up for change).