#13 Add multithreading to vision processing
This is done by utilizing an ExecutorService to handle
the execution and shutdown of vision threads. Since
each vision thread needs two separate network ports, and the range
is only of 10 ports (1180-1190), there is only a max of 5 vision threads.
It is unlikely that one will need that many threads anyway.